+++ packages/apollo-angular-link-http/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### vNEXT
 
+- Allow to not sending queries
+- Use same options in both, context and global
+
 ### v1.0.1
 
 - Serialize variables and extensions when using params ([PR #402](https://github.com/apollographql/apollo-angular/pull/402))